fix: adapt session loop command to current API types
